requirements.txt
streamlit==1.32.0
chromadb==0.4.22
python-dotenv==1.0.1
langchain==0.1.9
ollama==0.1.7
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.9
pypdf==4.0.2
docx2txt==0.8 